requirements.txt
google-generativeai==0.8.5
python-dotenv==1.1.0
requests==2.32.3
twilio==9.5.2
